+++ NEWS.md
@@ -2,6 +2,12 @@
 
 # handwriter 3.2.2
 
+## New features
+
+* Created `get_cluster_fill_rates()` to calculate the proportion of graphs assigned to each cluster in a cluster template.
+
+* Created `get_writer_profiles()` to estimate writer profiles for all handwritten documents in the input directory. Writer profiles are the cluster fill counts or the cluster fill rates. This function runs `process_batch_dir()` to split the handwritting int component shapes called graphs. Then, it runs `get_clusters_batch()` to assign each graph to a cluster in a cluster template. Lastly, it runs `get_cluster_fill_counts()` or `get_cluster_fill_rates()` depending on whether the user chooses counts or rates.
+
 ## Minor improvements and fixes
 
 * Changed the `writer_indices` and `doc_indices` arguments in `get_clusters_batch()` to be optional. The main reason for this change is so that users do not need to use a naming convention for their handwriting files in order to use `handwriterRF::calculate_slr()` and `handwriterRF::compare_documents()` which call `get_clusters_batch()`.